Changed computed expression type to int64_t
This allows the computer value for an expression to have a wide range of integer values

+++ UART/calculator.h
@@ -2,19 +2,19 @@
 #ifndef CALCULATOR_H_
 #define CALCULATOR_H_
 
-#include <stdlib.h>
+#include <stdint.h>
 
 char operatorStack[100];
 char* operatorSP;
 
-int operandStack[100];
-int* operandSP;
+int64_t operandStack[100];
+int64_t* operandSP;
 
 #define push(sp, n) (*((sp)++) = (n))
 #define pop(sp) (*--(sp))
 #define peek(sp) (*(sp - 1))
 
-int infixEval(char *infix, int length);
+int64_t infixEval(char *infix, int length);
 int getOperatorPrecedence(char op);
 
 #endif /* CALCULATOR_H_ */
